Extract interfaces for service classes
